+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKBPGCCVYAYY8RH35XPBSHQZ"
+        "01KKBQR2N68S3PJBJNRC4BAV37"
     )
     r"""The unique identifier of the data source"""