+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKBPGCCRAF62A0CCTCY0JRWB"
+        "01KKBQR2N1MZ7RW35QAPKXV71F"
     )
     r"""The unique identifier of the data source"""