+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKBPGCCXCBRYGM41FPRRYTPN"
+        "01KKBQR2NJ8XCMW9ARBR174QYX"
     )
     r"""The unique identifier of the data source"""